backend/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.0
Flask-JWT-Extended==4.5.0
Flask-Bcrypt==1.0.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
